+++ assets/javascripts/kitten/components/images/image-cropper/image-cropper.stories.js
@@ -9,3 +9,7 @@ export default {
 export const Default = args => {
   return <ImageCropper {...args} />
 }
+
+export const WithImage = () => {
+  return <ImageCropper imageSrc="/steven.jpg" fileName="Steven" />
+}