+++ assets/javascripts/kitten/components/steppers/vertical-stepper/components/title.js
@@ -29,7 +29,6 @@ export const Title = ({ variant, className, children }) => {
     >
       {children}
     </StyledTitle>
-
   )
 }