+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Add `variant` prop to `VerticalStepper`, with default value `andromeda`.
+- Feature: Add `DotIcon` component.
 
 ## [2.95.3] - 2020-11-05